backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
SQLAlchemy==2.0.23
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.1.0
python-dotenv==1.0.0
marshmallow==3.20.0
Werkzeug==3.0.1